refactor(rust cycle 19): admin_ops/user_handlers/crud 389→{list 79, get 29, update 136, delete 41, whoami 25}
