+++ src/bin/email_api_dir/admin_ops/user_handlers/crud_delete.rs
@@ -1,0 +1,41 @@
+#![allow(unused_imports, dead_code)]
+use super::super::*;
+use super::audit::log_admin_action;
+
+pub(crate) async fn api_admin_user_delete(
+    req: HttpRequest,
+    path: web::Path<String>,
+    mongo: web::Data<Arc<mongodb::Client>>,
+) -> impl Responder {
+    let actor = match admin_auth::require_admin(&req, &mongo, &mongo_db_name()).await {
+        Ok(u) => u,
+        Err(resp) => return resp,
+    };
+    let id = path.into_inner();
+    let coll = mongo
+        .database(&mongo_db_name())
+        .collection::<AdminUserRecord>(ADMIN_USERS_COLL);
+
+    match coll.delete_one(doc! { "id": &id }).await {
+        Ok(res) if res.deleted_count > 0 => {
+            log_admin_action(
+                mongo.as_ref(),
+                &actor,
+                "user.delete",
+                "admin_user",
+                &id,
+                None,
+                None,
+            )
+            .await;
+            HttpResponse::Ok().json(serde_json::json!({ "deleted": true, "id": id }))
+        }
+        Ok(_) => HttpResponse::NotFound()
+            .json(serde_json::json!({ "deleted": false, "message": "User not found" })),
+        Err(e) => {
+            eprintln!("api_admin_user_delete error: {}", e);
+            HttpResponse::InternalServerError()
+                .json(serde_json::json!({ "message": "Failed to delete user" }))
+        }
+    }
+}
