+++ src/bin/email_api_dir/admin_ops/user_handlers/crud_update.rs
@@ -1,0 +1,136 @@
+#![allow(unused_imports, dead_code)]
+use super::super::*;
+use super::audit::log_admin_action;
+
+pub(crate) async fn api_admin_user_patch(
+    req: HttpRequest,
+    path: web::Path<String>,
+    body: web::Json<UpdateAdminUserInput>,
+    mongo: web::Data<Arc<mongodb::Client>>,
+) -> impl Responder {
+    let actor = match admin_auth::require_admin(&req, &mongo, &mongo_db_name()).await {
+        Ok(u) => u,
+        Err(resp) => return resp,
+    };
+    let id = path.into_inner();
+    let coll = mongo
+        .database(&mongo_db_name())
+        .collection::<AdminUserRecord>(ADMIN_USERS_COLL);
+
+    let current = match coll.find_one(doc! { "id": &id }).await {
+        Ok(Some(v)) => v,
+        Ok(None) => {
+            return HttpResponse::NotFound()
+                .json(serde_json::json!({ "message": "User not found" }))
+        }
+        Err(e) => {
+            eprintln!("api_admin_user_patch read error: {}", e);
+            return HttpResponse::InternalServerError()
+                .json(serde_json::json!({ "message": "Failed to load user" }));
+        }
+    };
+
+    let now = now_iso();
+    let mut updated = current.clone();
+
+    if let Some(role) = &body.role {
+        let role = role.trim().to_ascii_lowercase();
+        if !["user", "admin", "support"].contains(&role.as_str()) {
+            return HttpResponse::BadRequest()
+                .json(serde_json::json!({ "message": "role must be user|admin|support" }));
+        }
+        updated.role = role;
+    }
+
+    if let Some(status) = &body.status {
+        let status = status.trim().to_ascii_lowercase();
+        if !["active", "restricted"].contains(&status.as_str()) {
+            return HttpResponse::BadRequest()
+                .json(serde_json::json!({ "message": "status must be active|restricted" }));
+        }
+        updated.status = status;
+    }
+
+    if let Some(email) = &body.email {
+        let email = email.trim();
+        if email.is_empty() || !email.contains('@') {
+            return HttpResponse::BadRequest()
+                .json(serde_json::json!({ "message": "invalid email" }));
+        }
+        updated.email = email.to_string();
+    }
+    if let Some(dn) = &body.display_name {
+        let dn = dn.trim();
+        updated.display_name = if dn.is_empty() {
+            None
+        } else {
+            Some(dn.to_string())
+        };
+    }
+    if let Some(v) = body.two_factor_enabled {
+        updated.two_factor_enabled = v;
+    }
+    if let Some(notes) = &body.notes {
+        let notes = notes.trim();
+        if notes.len() > 1024 {
+            return HttpResponse::BadRequest()
+                .json(serde_json::json!({ "message": "notes too long (max 1024)" }));
+        }
+        updated.notes = if notes.is_empty() {
+            None
+        } else {
+            Some(notes.to_string())
+        };
+    }
+
+    updated.updated_at = now.clone();
+    updated.last_activity_at = Some(now.clone());
+    updated.actions7d += 1;
+    let note = body
+        .role
+        .as_ref()
+        .map(|r| format!("Role changed to {}", r))
+        .unwrap_or_else(|| "User updated".to_string());
+    let mut recent = updated.recent_activity;
+    recent.insert(
+        0,
+        AdminUserActivity {
+            at: now,
+            label: note,
+            kind: "role_change".to_string(),
+        },
+    );
+    recent.truncate(8);
+    updated.recent_activity = recent;
+
+    match coll
+        .replace_one(doc! { "id": &id }, &updated)
+        .upsert(false)
+        .await
+    {
+        Ok(_) => {
+            log_admin_action(
+                mongo.as_ref(),
+                &actor,
+                "user.patch",
+                "admin_user",
+                &id,
+                None,
+                Some(serde_json::json!({
+                    "role": updated.role,
+                    "status": updated.status,
+                    "email": updated.email,
+                    "displayName": updated.display_name,
+                    "twoFactorEnabled": updated.two_factor_enabled,
+                })),
+            )
+            .await;
+            HttpResponse::Ok().json(serde_json::json!({ "user": updated }))
+        }
+        Err(e) => {
+            eprintln!("api_admin_user_patch write error: {}", e);
+            HttpResponse::InternalServerError()
+                .json(serde_json::json!({ "message": "Failed to update user" }))
+        }
+    }
+}
