+++ src/bin/email_api_dir/admin_ops/user_handlers/mod.rs
@@ -1,7 +1,17 @@
 pub(crate) mod audit;
-pub(crate) mod crud;
+pub(crate) mod crud_create;
+pub(crate) mod crud_delete;
+pub(crate) mod crud_get;
+pub(crate) mod crud_list;
+pub(crate) mod crud_update;
+pub(crate) mod crud_whoami;
 pub(crate) mod lifecycle;
 
 pub(crate) use audit::*;
-pub(crate) use crud::*;
+pub(crate) use crud_create::*;
+pub(crate) use crud_delete::*;
+pub(crate) use crud_get::*;
+pub(crate) use crud_list::*;
+pub(crate) use crud_update::*;
+pub(crate) use crud_whoami::*;
 pub(crate) use lifecycle::*;
